requirements.txt
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
openai>=1.3.0